requirements.txt
Babel==2.12.1
bcrypt==4.0.1
blinker==1.5
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mysqlclient==2.1.1
packaging==23.0
passlib==1.7.4
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
